Split conditions to own files
